Create new channels for each peer message. (#43)
When dcrseeder is contacting a set of potential peers, channels are used to block until  verack/addr messages have been received.

A single channel  was being used for all verack messages from all clients, and another single channel for all addr messages from all clients. This is problematic for obvious reasons - the verack/addr messages are practically guaranteed to arrive out of order, so when a value is received on a channel, the code has no way of knowing which peer the message has come from.

With this PR, a channel is created for every message. As a side effect, the peer.Config struct is also being recreated for each message - idk if that is a problem?