Remove quit chan from Manager.
Make the main loop of the Manager a blocking func which accepts a
context. Shutdown the Manager when the context is canceled.

This slightly alters the semantics of the code but in a way that has
functionally no impact. Previously the Tickers which update the Manager
would only be stopped after the HTTP server and creeper had stopped, but
now they are stopped immediately when the context is canceled.